css mypage (#53)
* Update aws version in buildspec.yml (#20)

* feature3

* feature3

* feature3

* Merge branch 'dev' of codestates/shelp into feature3

* Login/out signin/out errhandling

* Css mypage

+++ shelp-client/src/pages/mypage.js
@@ -8,7 +8,7 @@ const serverUrl = "http://localhost:4000";
 const Container = styled.div`
   display: flex;
   flex-direction: column;
-  position: absolute;
+  /* position: absolute; */
   top: 0;
   left: 0;
   width: 100%;
@@ -30,9 +30,10 @@ const Profile = styled.div`
   flex: 0.1 0 auto;
   background-color: white;
   display: flex;
+  flex-direction: column;
   align-items: center;
 
-  > img.profile-left {
+  /* > img.profile-left {
     position: relative;
     width: auto;
     height: 70%;
@@ -41,16 +42,25 @@ const Profile = styled.div`
     padding-top: 1rem;
     overflow: hidden;
     background-color: white;
-  }
+  } */
 
-  > div.profile-right {
+  /* > div.profile-right {
     flex: 1 0 auto;
     top: 50%;
     height: 70%;
     margin: 10vh;
     display: flex;
-    flex-direction: column;
-
+    flex-direction: column; */
+    > div.profile-image {
+      border-radius: 70%;
+      overflow: hidden;
+      height: 300px;
+      margin-top: 30px;
+      box-shadow: 0em 0em 1em rgba(0, 0, 0, 3);
+    }
+    > div.profile-container {
+      margin-top: 30px;
+    }
     > div.profile-email {
       height: auto;
 
@@ -60,44 +70,119 @@ const Profile = styled.div`
       height: auto;
       display: flex;
       padding: 1em 0;
-
-      > div.name {
-        width: auto;
-        padding-right: 0.5em;
-        margin-top: auto;
-        margin-bottom: auto;
-        font-size: 4em;
-      }
-      > div.welcome-message {
-        margin-top: auto;
-        margin-bottom: auto;
-        font-size: 2em;
-      }
+    }
+    > div.name {
+      width: auto;
+      padding-right: 0.5em;
+      margin-top: auto;
+      margin-bottom: auto;
+      font-size: 4em;
+    }
+    > div.welcome-message {
+      margin-top: auto;
+      margin-bottom: auto;
+      font-size: 2em;
     }
     > div.profile-desc {
       color: rgba(0, 0, 0, 0.4);
       height: auto;
       font-size: 1.5em;
     }
-  }
+  
 `;
+const User = styled.div`
+  height: 150px;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: space-between;
 
+  > div.periodButton {
+    display: flex;
+    flex-direction: row;
+    justify-content: space-between;
+    width: 200px;
+    
+    button {
+      border: 3px solid gray;
+      border-radius: 10px;
+    }
+    > button.alerm{
+      border-style: solid;
+      border-color: red;
+    }
+  } 
+`
+const EditComp = styled.div`
+  height: 150px;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: space-between;
+
+  div {
+    input {
+      border-style: solid;
+      border-radius: 10px;
+      border-color: gray;
+      height: 25px;
+      width: 200px;
+    }
+  }
+
+  > div.periodButton {
+    display: flex;
+    flex-direction: row;
+    justify-content: space-between;
+    width: 200px;
+
+    > button {
+      border: 3px solid gray;
+      border-radius: 10px;
+    }
+
+    > button.clicked {
+      border-color: red;
+    }
+  }
+`
 const Collection = styled.div`
+  display: flex;
+  text-align: center;
+  justify-content: center;
+  height: 100px;
   flex: 10 0 auto;
   background-color: white;
   margin: 0 2rem 2rem 2rem;
-  box-shadow: 0em 0em 1em rgba(0, 0, 0, 0.3);
+  /* box-shadow: 0em 0em 1em rgba(0, 0, 0, 0.3); */
   border-radius: 0.5rem;
+  margin-top: 70px;
+
+  > div.buttonBox {
+    display: flex;
+    flex-direction: column;
+    justify-content: space-evenly;
+    /* align-items: center; */
+
+    button {
+      width: 300px;
+      height: 30px;
+      margin-top: auto;
+    }
+  }
 `;
 
 const AlarmSet = styled.div``;
 
 function Mypage({ isLogin, userinfo, setUserinfo }) {
+  // console.log(userinfo.period)
   const [button, setButton] = useState(true);
   const [per, setPer] = useState(1);
   const [img, setImg] = useState();
 
+
   const handleChange = (e, key) => {
+    console.log(e.target.value)
     setUserinfo({ ...userinfo, [key]: e.target.value });
   };
 
@@ -106,29 +191,36 @@ function Mypage({ isLogin, userinfo, setUserinfo }) {
   };
 
   const handleButtonSave = () => {
-    const formData = new FormData();
-    formData.append("img", img);
-    // const { name, password, image, desc } = userinfo;
-    setUserinfo({ ...userinfo, image: formData });
+    setUserinfo({ ...userinfo });
     axios.put(`${serverUrl}/profile`, userinfo).then((res) => {
-      console.log(res);
-    });
-
+      if(res.status === 200){
+        alert ("저장되었습니다.")
+      }
+    })
     setButton(!button);
   };
 
   const deleteUser = () => {
-    axios.delete(`${serverUrl}/profile`).then((res) => {
-      console.log(res);
-      window.localStorage.clear();
-      window.location.replace("/");
-    });
+    const message = window.confirm('정말 탈퇴하시겠습니까?');
+    if(message){
+      axios.delete(`${serverUrl}/profile`).then((res) => {
+        console.log(res);
+        window.localStorage.clear();
+        console.log('word')
+        window.location.replace("/login");//////////////////////////////////////////////////////////////////////////////////////////////////
+      });
+    }else{
+
+    }
   };
 
-  const loadFile = (e) => {
-    // setUserinfo({...userinfo, image: e.target.files[0]});
-    setImg(e.target.files[0]);
-  };
+  const temp = () => {
+    console.log('hellof')
+    console.log(userinfo.period)
+    if(userinfo.period === '5'){
+      console.log('clicked')
+    }
+  }
 
   return (
     <Container>
@@ -140,15 +232,21 @@ function Mypage({ isLogin, userinfo, setUserinfo }) {
           </div>
           <div className="profile-container">
             {button ? (
-              <div>
-                <div className="profile-name">{userinfo.name}</div>
-                <span>님 환영합니다.</span>
+              <User>
+                <div className="profile-name">{userinfo.name}님 환영합니다.</div>
+                {/* <span>님 환영합니다.</span> */}
                 <div className="profile-email">{userinfo.email}</div>
-                <div>{userinfo.period}</div>
+                {/* <div>{userinfo.period}</div> */}
+                <div className='periodButton'>
+                  <button onClick={(e)=>{handleChange(e, "period")}} value='1' className={userinfo.period==='1'? 'alerm': ''}>1d</button>
+                  <button onClick={(e)=>{handleChange(e, "period")}} value='3' className={userinfo.period==='3'? 'alerm': ''}>3d</button>
+                  <button onClick={(e)=>{handleChange(e, "period")}} value='5' className={userinfo.period==='5'? 'alerm': ''}>5d</button>
+                  {/* <div></div> */}
+                </div>
                 <div>{userinfo.desc}</div>
-              </div>
+              </User>
             ) : (
-              <div>
+              <EditComp className="editComp">
                 <div>
                   <input
                     type="text"
@@ -165,12 +263,11 @@ function Mypage({ isLogin, userinfo, setUserinfo }) {
                   />
                   로 변경
                 </div>
-                <div>
-                  <input
-                    type="number"
-                    onChange={(e) => handleChange(e, "period")}
-                  />
-                  로 변경
+                <div className='periodButton'>
+                  <button onClick={(e)=>{handleChange(e, "period")}} value='1'>1d</button>
+                  <button onClick={(e)=>{handleChange(e, "period")}} value='3'>3d</button>
+                  <button onClick={(e)=>{handleChange(e, "period")}} value='5'>5d</button>
+                  <div></div>
                 </div>
                 <div>
                   <input
@@ -180,13 +277,13 @@ function Mypage({ isLogin, userinfo, setUserinfo }) {
                   />
                   로 변경
                 </div>
-              </div>
+              </EditComp>
             )}
           </div>
         </Profile>
         {/* <Collection> */}
         <Collection>
-          <div>
+          <div className="buttonBox">
             {button ? (
               <button onClick={handleButtonEdit}>edit</button>
             ) : (
